+++ nand_chip/nand_drive.h
@@ -5,21 +5,21 @@ using NandDrivePtr = std::shared_ptr<NandDrive>;
 
 class NandDrive {
 public:
-    static constexpr uint32_t DIE_CNT    = 1;
-    static constexpr uint32_t PLANE_CNT  = 1;
-    static constexpr uint32_t BLOCK_CNT  = 64;
-    static constexpr uint32_t PAGE_CNT   = 8;
-    static constexpr uint32_t PG_SZ      = 16384;
+    static constexpr uint64_t DIE_CNT    = 1;
+    static constexpr uint64_t PLANE_CNT  = 1;
+    static constexpr uint64_t BLOCK_CNT  = 64;
+    static constexpr uint64_t PAGE_CNT   = 8;
+    static constexpr uint64_t PG_SZ      = 16384;
     static constexpr size_t   PAGE_TOTAL_BYTES = PG_SZ;
     static constexpr size_t   TOTAL_PAGES = static_cast<size_t>(DIE_CNT) * PLANE_CNT * BLOCK_CNT * PAGE_CNT;
     static constexpr size_t   TOTAL_BYTES = TOTAL_PAGES * PAGE_TOTAL_BYTES;
     static constexpr size_t   TOTAL_USER_BYTES = TOTAL_PAGES * PG_SZ;
 
     struct Addr {
-        uint32_t die;
-        uint32_t plane;
-        uint32_t block;
-        uint32_t page;
+        uint64_t die;
+        uint64_t plane;
+        uint64_t block;
+        uint64_t page;
     };
 
     NandDrive();
