+++ transaction/transaction.h
@@ -14,21 +14,21 @@ using TransactionErasePtr = std::shared_ptr<TransactionErase>;
 
 class PhysicalPageAddress{
 public:
-    uint32_t channel_id;
-    uint32_t chip_id;
-    uint32_t die_id;
-    uint32_t plane_id;
+    uint64_t channel_id;
+    uint64_t chip_id;
+    uint64_t die_id;
+    uint64_t plane_id;
     uint64_t block_id;
     uint64_t page_id;
-    PhysicalPageAddress(const uint32_t channel_id=0, const uint32_t chip_id=0, const uint32_t die_id=0,
-                        const uint32_t plane_id=0, const uint64_t block_id=0, const uint64_t page_id=0)
+    PhysicalPageAddress(const uint64_t channel_id=0, const uint64_t chip_id=0, const uint64_t die_id=0,
+                        const uint64_t plane_id=0, const uint64_t block_id=0, const uint64_t page_id=0)
         : channel_id(channel_id), chip_id(chip_id), die_id(die_id), plane_id(plane_id), block_id(block_id), page_id(page_id) {}
     PhysicalPageAddress(const PhysicalPageAddress& other) = default;
 };
 class Transaction {
 public:
     uint64_t transaction_id;
-    uint32_t stream_id;
+    uint64_t stream_id;
     TransactionType type;
     TransactionSourceType source;
     Priority priority;
