+++ user_request.h
@@ -3,12 +3,12 @@ enum class UserRequestType { READ, WRITE, TRIM };
 
 class UserRequest {
 public:
-    uint32_t id;
+    uint64_t id;
     UserRequestType req_type;
     uint64_t start_lpa;
     uint64_t size_in_sectors;
     uint64_t size_in_bytes;
-    uint32_t stream_id;
+    uint64_t stream_id;
     std::vector<uint8_t> data;
     uint64_t sectors_from_cache;
 };
